+++ frontend/src/components/Scoreboard/Round.tsx
@@ -11,7 +11,7 @@ import {
 } from "@mui/material";
 import { enqueueSnackbar } from "notistack";
 import { useEffect, useState } from "react";
-import { Scoreboard } from "../../models/Scoreboard";
+import { Scoreboard } from "../../models/Scoreboard/Scoreboard";
 import { Round } from "../../models/ent";
 import DoubleArrowLeftIcon from "@mui/icons-material/KeyboardDoubleArrowLeft";
 import DoubleArrowRightIcon from "@mui/icons-material/KeyboardDoubleArrowRight";
