+++ frontend/src/models/Scoreboard/Check.tsx
