+++ frontend/src/models/Scoreboard/Status.tsx
