+++ frontend/src/models/Scoreboard/Team.tsx
